dtu-test-progs: fixed bug due to virt->phys workaround.
Because we use SE mode currently, the programs run in virtual
memory and the DTU translates that to physical memory. However,
the workaround assumes that receive buffers and messages are
contiguous in physical memory.
To avoid a more complex workaround in the DTU I have decided
to temporary fix that by using page-aligned memory in the
test programs and ensure that no message or buffer is larger
than a page.