Replaced debugSymbolTable with a symtab per process/system.
Since we load different programs on different PEs, this is an
essential feature for us. Note however, that the change is not
yet complete. There is still something to do for ARM, SPARC and
Alpha. Basically, the debugSymbolTable is still there and filled
on these architectures, but not used anymore.