DTU: put a specified "memory-command" on unused PEs.
These are used as memory-PEs, where the core should do nothing
and the scratchpad can be accessed by others.
Of course it would be nicer to not have a core at all then. The
problem is that each System keeps some arrays for statistics that
are large enough to be indexed by every master in that System.
As we have a System for each PE, and sometimes these statistics
are accessed by masters from other Systems (maybe that's wrong
anyway), all arrays have to be large enough to be indexed by
every master in the whole tree. So, an easy solution is to just
make all PEs the same.