DTU: no core for mem-PEs and change masterId for SPM accesses.
Having a core for memory-PEs is not a good idea. The scratchpad
gets filled with all kinds of things (GDT, IDT, code, stack,
...). That makes it difficult to initialize the scratchpad with
a file.
Thus, we go a different path now. We fix the problem with the
masterId by setting the masterId of the DTU when accessing the
scratchpad. Thus, it's always a local masterId, so that the
problem does not occur. We needed to add a setter for the
masterId though, because changing it was not indended as it
seems. But it works :)