DTU: fixed race condition when receiving messages.
We have to increment the write-pointer directly when receiving
a message. If doing that later, another message can arrive,
which will then overwrite the last one.

I've decided that we don't need to do that in a separate cycle,
because we have very few dependencies. And the registers that
are read overlap quite a bit.