DTU: allow suspending the core while waiting for messages.
This works as follows: the DTU has a global message-count
register, that contains the number of unprocessed messages.
It has to be in sync with the sum of all BUF_MSG_CNT registers
and will be updated by the DTU automatically.

The software can thus check if there are messages pending and
if not, suspend the core. The DTU will wake up the core as soon
as a message has arrived.

To resolve the race condition between checking the pending
messages and suspending the core, I've introduced a "pin" at
the core that is set/unset the by DTU according to whether
messages are pending. This pin will prevent the core from
sleeping if messages are pending. In gem5, it is checked
before going to sleep and if there are messages, the core
stays active.