DTU: changed the register layout.
In theory, if the DTU was built that way, we could have executed
multiple commands in parallel, e.g. doing an asynchronous read
and sending a message. However, it makes the DTU more complex
than necessary and I don't think that there are many usecases
for parallel commands that really give an performance advantage.

Therefore, I have decided that the endpoint registers contain
only configuration. Registers like MSG_ADDR, MSG_SIZE, ... are
now command registers. This means that we have much less
endpoint registers, which allows to have more endpoints and makes
save/restore of endpoints faster. On the other hand, there will
always be just one command in parallel.

As we want to implement permissions soon, I have put STATUS and
MSG_CNT into a separate register file, i.e. Dtu registers. They
are readonly in general. The command registers are writable by
everyone, while the endpoint registers are only writable by
privileged PEs.

Note also that I have updated the dtutest accordingly. But atm,
it just compiles, but does not work. I'm not sure if it makes
sense to maintain the test anyway. But it might be a good
starting point if we want to add unittests later that stress
the DTU in a way that SW can't.