DTU: replaced EP mode register with command opcode.
That is, now the operation (send, reply, read, write) is
specified by writing to the command register. In this way,
we need one EP register less and it better suites our
needs. Because at least for reading/writing, an endpoint
can (sometimes) be used for both. At the same time, the
application should not be able to change the mode. But it
had to to specify what to do. So, this problem is gone now.
The application does not need to write to any EP register
anymore.