DTU: introduced PRIV flag and addd enforcement of permissions.
That is, we have a PRIV flag in the STATUS register now that
indicates whether the DTU belongs to a privileged PE. At boot,
all PEs are privileged. The kernel can than clear the PRIV
flag at all other PEs to make them non-privileged.

Only if PRIV is set, the PRIV flag and EP registers can be
changed.