DTU: don't separate command and busy-flag.
The problem is that if writing to the command register starts
a command and reading from the status register checks when it's
done, that the compiler does not see a connection there. So, it
assumes, by default, that he can reorder that.
To prevent that problem, the DTU is now busy if command is not
IDLE, so that we write and read at the same memory location.