Workaround for suspend with O3 CPU.
It seems that the O3 CPU does not continue to tick after
executing hlt, no matter whether we stop in suspend()
or not. I don't know what exactly is going on. But we can
fix that by waking the core up again.
Performance-wise it does not matter, because we will only
do that if we had bad luck and between checking whether
messages are pending and doing hlt, a message arrived.