DTU: relaxed alignment requirements and send synchronously.
That is, although we still expect the header to be 32 bytes large,
the block size to be 32 bytes and aligned receive buffers, the address
at the initiator does not need to be aligned anymore.

Furthermore, we now only send out the next data packet if we have
received a response for the last. Since the crossbar has a limited
packet queue and the DMA controller does it in the same way, it
seems to be better for us, too.