DTU: introduced local buffers to send/receive.
Splitting a packet in pieces which actually belong together
seems to be a bad idea in gem5. Because the model seems to be
that you always send a packet in one piece and the NoC handles
that appropriately (although it is actually transferred at
once).
This is especially a problem when talking to DRAM, because the
performance difference between sending many small packets and
one large one is tremendous.
Thus, we now send everything in one packet again. To make that
work, I introduced local buffers in the DTU, in which packets
that should be sent are collected and data that should be
received is parked first. The XferUnit performs the exchange of
data between the local buffer and the cache.