DTU: handle larger requests than buf_size.
In theory, we could send larger packets than buf_size, even if
we're sending to a cache-PE. Because we could use the buffer
as a ring and thereby handle that packet piece by piece.
But as the idea in gem5 seems to be to send a packet at once,
we can't do that because we need to place the packet in the
buffer in one go (or at least it would get complicated to do it
differently.
Therefore, we limit the packet size to buf_size globally and
simply issue multiple read/write requests, if necessary. For
messaging, we assume that a message fits in the buffer.