DTU: use large buffers for SPM/DRAM-PEs.
For memory PEs or PEs with SPM, we do not need a buffer. For the sake
of an easy implementation, we just make the buffer very large and
the block size as well, so that we can read a packet from SPM/DRAM
into the buffer and send it from there. Since that costs no simulated
time it is the same as having no buffer.