DTU: connect the cache to a memory-PE.
That is, now we have a memory-EP that is preconfigured to
access a specified range in some memory-PE. The memory
port of the cache is connected again to DTU, which looks
into the memory-EP and sends a corresponding request to
the memory-PE.
I have also changed the way we handle invalid speculative
memory accesses. Previously, we catched them before going
to the cache/SPM. Now, we let the SPM catch all accesses
and ignore them. And if a cache is used, we handle that
when checking the memory-EP. So, we handle it at the places
where we directly know whether it's invalid or not.