DTU: release the noc-port if we passed the request to XferUnit.
That is, we no longer make it busy until we send the response,
but when we passed it to the XferUnit. Because this will handle
parallel requests itself and we no longer need the "path that the
port triggers".
Especially for DRAM this is a good thing to do, because each
request takes a considerable amount of time and the DRAM
controller might be able to batch them with the new approach.