DTU: changed EP register layout.
I have now reduced the number of EP registers to the minimum by
introducing a type-field, so that we know what kind of EP it is,
reusing the bits for different types and moving things closer
together. With this, we have just 3 64-bit values per EP, not 14
as before.
This reduces the time to configure endpoints, invalidate them,
context-switch them and so on.
Debuggability was not reduced with this, because the type-field
enables us to decode the bits correspondingly and thus display
to the user the changes to EP registers.
Note that I had to disable things in the dtutester again.