DTU: continue on pagefaults.
Atm, we assume that no application causes a pagefault
accidentally, i.e. because of a bug. If it occurs
during speculation, this is ok, because the CPU won't
use the result.
At some point, we want of course to notice accidental
pagefaults. We probably need to change the CPU for that
a bit to react on errors for memory accesses, because
atm it doesn't do that.