DTU: added INTERN flag to PTEs and handle all PFs.
The INTERN flag is required for the core to access a page.
For external requests, it is not required. Since only the
kernel hands out memory capabilities to do external requests,
he is allowed to access pages that don't have the INTERN flag
set, but others are not. With this, we protect PTEs from being
read/written by the core itself.
Furthermore, we handle pagefaults at all places now.