DTU: added external command for cache invalidation.
The invalidation of caches is required when a VPE is
destroyed to ensure that the next VPE running on the PE
has a fresh cache and not some memory cached that is
reused by the OS for other purposes.

Currently, we add all caches to the DTU and make the Cache
object known to it. This way, we can directly call
memInvalidate() on all caches.

This is of course a hack. The good way of implementing that
would be to introduce a new request type, send it to the cache,
which should then invalidate all cache lines and send it to the
next cache in the hierarchy until we reach the DTU again.

Our current solution has the advantage though, that we do not
(really) need to change the cache. We need to make a few methods
public that are protected by default, but we don't need to change
the internal logic.

Note also that we currently not simulating a delay for
invalidating all cache lines. But this will change in a follow-up
commit, when we support delays for external commands.