DTU: added boot modules and map only required parts.
The M3X86System provides now boot modules on the kernel
PE and we put idle on all application PEs. This enables
the kernel to load the boot modules on the application
PEs as desired.

Furthermore, we don't map a large part of the address
space on the kernel PE, but only the required parts.
We do still map a large part on the application PEs,
because M3 cannot handle non-paged VPEs properly yet.