DTU: added a read-write barrier and the enforcement.
This is used by the OS to ensure that receive buffers are
in readonly memory. The DTU provides a register that defines
the barrier between writeable space and readable space. This can
thus be configured for each PE individually.
We do that with a barrier to support that with virtual memory
and without.