DTU: bugfix in pagefault handling.
We actually can't merge a PF with pending PFs. Because with
this, we reorder requests. E.g., if a read request comes first
and issues a pagefault, then a write request follows and then
again a read request. In this case, we would merge the last
read request with the already existing one and with this, we
consider it finished before the write request. So, it might
get the wrong data.

Thus, I've changed that now to no longer merge request, but
just keep a queue of them and check the TLB again before
we send a new pagefault. This is worse performance-wise,
because for unmapped things (due to speculation) we send
multiple pagefaults. But this has to be improved in a
different way.