DTU: moved vpeId out of NocAddr into senderState.
It was a bad idea to store that in NocAddr, because we
don't have that many bits left there and we might want
to have many VPEs. Furthermore, increasing the size of
NocAddr increases caches since they are stored there.

Instead, we store it in the senderState, that we already
have for all transfers. So, in the header, kind of.

The consequence is that we can't map virtual address
into our address space. But this was a bad idea for
other reasons as well.