DTU: introduced NOPF flag in command register.
That allows the user to do a memory transfer and abort it,
if a pagefault occurs. This is useful for the kernel and the
pager, who are involved in pagefault handling. They definitely
don't want to cause a pagefault at the paged application again.

Furthermore, we need that to ensure that we can always resolve
pagefaults eventually. Because we need a special transfer buffer
that is only used if no pagefaults can occur. That is, if the
NOPF flag is given, we use this buffer, otherwise only the other
ones. This way, we know that we can always use this buffer to
resolve pagefaults.

In this way, we handle the case where a transfer can potentially
cause another transfer on the same PE (indirectly). Because e.g.
the kernel needs to read/write PTEs.