DTU: added a simple hash accelerator.
The accelerator can be used instead of a core and implements a
simple state machine, that starts when a message has been received,
reads the message, computes the hash from a part of the memory,
belonging to the sender (identified via label) and sends the
reply.

For hashing, SHA* are supported, for which openssl is used.
Unfortunately, it seems to be necessary to modify the SConstruct
file, because adding the libraries in the SConscript has no
effect.