DTU: adding basic time features.
We now have a SLEEP command that suspends the core for a given number
of cycles or until we get waked up (message receive, IRQ, ...). For
that, the command handling has been changed to send the response for
command register writes as soon as the command has finished executing.
This will implicitly put the core into a wait-for-memory state until a
command has been executed, which is especially nice for the sleep cmd.

Furthermore, we have two new registers: CUR_TIME and IDLE_TIME. The
former does always return the current time in cycles and the latter
contains the sum of all sleep durations.