DTU: replaced ringbuffer with msg array.
That is, we now use a EP register to keep track of which slots in the
array are occupied and which slots are unread. That allows the SW to
keep messages in the array even if it wants to handle other messages
before replying.

The obvious downside is that we have now limited a receive buffer to 32
messages. But this is acceptable, I think, because the SW can just use
multiple threads, which is a good idea if many clients should be served
anyway.

The DTU does now provide a FETCH_MSG command, which checks whether there
are unread messages and returns the first one. This is done in a round-
robin fashion to ensure that every request gets handled eventually.