DTU: temporary fix for delayed IRQ injections.
The O3 CPU takes some time to actually jump to the IRQ handler
and continues to execute quite a few instructions in the
meantime. This is a problem if it issues the SLEEP command
within these instructions. This commit fixes by ignoring
the sleep command between IRQ injections and aborts. This
is no longterm solution, though.