DTU: handle cache coherence properly.
We cannot bypass the CC protocol with read/write transfers. To adhere
to the protocol, we do now send memory requests that should be sent
to a memory PE into our local cache instead. The cache will handle it
correspondingly. This seems to be bad for performance, though, because
we are doing everything cache line by cache line. Using a prefetcher
helps a bit, but not that much, unfortunately.