DTU: use feature-flag to disable communication.
That is, instead of invalidating the VPE id, we set a flag in
a register to prevent further incoming communication with unprivileged
PEs. The kernel still needs to be able to communicate with the PE.

This allows us to still execute commands after we have aborted a
command, which is required to save the SPM content, for example.