DtuHashAccel: support failed replies.
That is, we handle the case now that the calling VPE might be
suspended if we try to reply. This is done really simple currently,
because we just wait until the kernel tells us that we can retry it.
We could make that more complicated by using the asynchronous version
of the activatereply syscall, saving the state and handling other
clients in the meantime.