DTU: allow safe forwarding of send/reply.
For that, the DTU "disables" a SendEP if a send failed because
the receiver is suspended by setting the target VPE ID to invalid.
Similarly, a flag in the message header is set if a reply failed
for the same reason. This prevents that someone can use the forward
feature in cases where the send/reply did not actually fail.