DTU: handle CPU requests properly.
We need to make sure that we send the requests to the cache in the
same order as we received them from the CPU. That means, if a
translation or pagefault handling is necessary, subsequent requests can
not be send to the cache earlier. At least not if they refer to the same
page. To achieve that, the DTU maintains a queue of translations and
dequeues them in FIFO order.

It would be enough to enqueue TLB hits as soon as one translation is
pending. However, we can improve that by allowing hits to be sent to the
cache, if it refers to a different page. To do so, the TLB stores the
number of ongoing translations. The difference seems to be measurable
only during warmup and if context switching is involved. I'm not sure
if a tagged TLB is enough to reduce the overhead during context
switching, so that this trick is actually not necessary.