DTU: fix the ordering problem properly.
We can't call finish_translate as soon as the translate is finished,
but have to wait until the CPU request is out of the queue. As this is
only about maintaining the expected order for CPU requests, we don't
have to do anything for other translations (e.g., from the XferUnit).