DtuAccel: added state and save/restore it.
That is, we now have a stateful API consisting of start, update
and finish to generate hashes for large amounts of data. This
requires us to save and restore the state during context switches.
We simply write that state including the buffer to external memory
via a memory EP, if we have no virtual memory.