Dtu: simplified the abort command.
Previously, we sent messages between DTUs to abort running commands.
This turned out to be quite difficult and only necessary, because we
assumed previously, that message reception can pagefault.

I think, the better trade-off is to keep the hardware simple and let
the SW make sure that message reception does not pagefault. This allows
us to simply wait if messages are in sent/received and forget
reads/writes that are in progress.