DTU: improved sleep/wakeup/irq handling.
I removed the INJECT_IRQ command to always use the write to the
EXT_REQ register. The SW is required to acknowledge the interrupt
by writing a 0 to EXT_REQ.

This required a change to the register permissions, because everybody
needs to be able to write to EXT_REQ. We now have again 2 privileged
levels: privileged and unprivileged. The request registers to handle
TLB misses in SW are now always accessible, but only have an effect if
the DTU has no PT walker. The kernel will map these registers
unaccessible for unprivileged SW.

I also removed the requirement for wakeup/suspend in the connector. That
is, the CUs do no longer need to support that. The SLEEP command of the
DTU will simply delay the response of the write request until we wake up
again. Note that we still use wakeup/suspend to improve the simulation
speed.

Change-Id: I