DTU: fixed bug in setting of IRQ_WAKEUP.
The FEATURES register needs to reflect the type of wakeup
after every wakeup. That is, we cannot just set it if we
were sleeping, but also need to unset it in the other case.
This is because we might be woken up first during a sleep
and then receive a message and start working and then be
woken up again.