DTU: use acknowledged events to ensure none is lost.
Previously, for example EP invalidation could easily happen unnoticed
by the core, because the core is just waked up, which has no effect if
the core is already awake.
Now we have events (EP invalid, credits received, message received) that
are stored as a bitmask in the global EVENTS register. If there are any
pending events, the DTU refuses to suspend the core. The core can ack
the events in an cmpxchg manner.
Received messages cannot be acked in this way though, because they need
to be acked already and should also be acked individually.c

Change-Id: Ie4c52de24c8e733d5ecd3ffce9de43d6d5458c6f