DTU: fixed bug in msg receive for foreign VPEs.
We don't need to raise a core request for a foreign-VPE receive when
starting the transfer. Instead we NEED to raise a core request in
that case if we want to finish the transfer, because at this point we
need to update the CUR_VPE register, which will only be done if the
receiving VPE is the current one.

The previous approach led to the problem that the receiver could run
when starting to receive a message, but someone else could run when
finishing the receive. In that case, we silently received the message
without incrementing the msg count in CUR_VPE AND without telling
PEMux about it. Thus, some time later the message count in CUR_VPE
was off, leading to an overflow in CUR_VPE when fetching a message,
and so on.