arch-riscv,arch-x86: moved common code into generic walker class.
The basic structure of the page table walker, that is, loading PTEs in
multiple steps until the leaf PTE is found, support multiple walks at
the same time, handling sends and receives of packets, supporting timing
and functional walks, etc. is generic. Therefore, I've moved this code
into the generic ISA as a base class for the RISC-V and X86 page table
walker. Unfortunately, the ARM PT walker is very different, so that it
does not take advantage of the generic PT walker.

The ISA-specific page table walker only needs to implement the setup
of the walk, each step of the walk, and a couple of helper methods.

Change-Id: I45b7e8a365b32b49f1c34a5501f6b4b0d015d74e