DTU: fixed problems in external command INV_EP.
Previously, we always updated the message count in CUR_VPE when
invalidating a receive EP. However, this does only work if the receive
EP belongs to the current VPE. Unfortunately, we cannot use the coreReq
mechanism to inform PEMux about it in case the EP does not belong to the
current VPE, because a coreReq always belongs to a transfer.

Therefore I decided to invalidate the EP, but report the unread mask to
software via the EXT_CMD register. Thus, if the kernel invalidates a
receive EP, it can check whether there were unread messages and if so,
tell PEMux about it, which will in turn update the VPE's message count.

Note that the receive EP is invalidated immediately to avoid that
further messages can arrive. This has the disadvantage that the message
count in CUR_VPE is temporarily too high if the receive EP belonged to
the current VPE. However, in the worst case, it keeps the VPE busy
checking for messages until the message count has been corrected.