TCU: removed credit-only-reply feature.
The problem is that it requires us to introduce a special case for core
requests in case the VPE that receives the credits is not running. That
is, setting the event in CUR_VPE cannot be done in that case and we need
to use a core request to inform PEMux. However, core requests are
coupled with transfers and there is no transfer.

This feature existed, because it reduces the number of required EPs,
because if the sender only wants to get its credits back and no real
reply, it doesn't need a receive EP on his side. Since we plan to
support "unlimited" EPs anyway, I don't think that the optimization
makes that much sense. Also because it was always tricky to think about
that case.