arch-riscv: use fp_enable_check and make FPU dirty on use.
Previously, we only checked whether the FPU is turned off and raised an
exception if so. Now we additionally make the FPU dirty to tell the OS
whether there is FPU state to save on a context switch. For simplicity,
we currently do not distinguish between read and write access, but
consider the FPU dirty on every access.

As for other ISAs, we now use fp_enable_check to reduce the code
duplication.

Change-Id: Ife14b0eeb5ff0afc75142ae4f7e73d61bbbbef1c