TCU: ensure that no interrupt gets lost.
This seems to be an x86-only problem, but since we have the
CLEAR_IRQ-ack on ARM and RISC-V already, it doesn't hurt to use this
approach in general. The idea is that the connector keeps a queue of
pending interrupt requests and only fires one at a time. The write to
CLEAR_IRQ is now mandatory on all ISAs and acknowledges the reception of
an IRQ and fires the next one, if any.