TCU: disallow PFs on remote accesses.
Pagefaults during remote accesses cause problems, because they can take
arbitrary amounts of time and thus we need to be able to abort them.
However, a distributed abort protocol is difficult and everything else
doesn't work properly (we need to ignore future responses, requiring
unique IDs, and we run out of xfer buffers at some point).

Thus, I've decided now to always disallow PFs on remote accesses. That
means, in case of a PF, the transfer is aborted and fails. So, the
software has to map and pin the memory to make sure that this does not
happen.

For that reason, I've removed the command flags.