TCU: removed xfer-buffer id from translation core request.
Now that pagefaults can never happen for remote transfers, PEMux knows
that a pagefault always results from a command. Thus, it doesn't need
to known which xfer buffer was used to check whether it's the same as
the command it aborted.

For that reason, we don't need to expose xfer buffers and their ids to
software at all.