TCU: reworked abort concept.
Now that we don't allow pagefaults on remote accesses anymore, abort can
be simplified: local transfers are aborted right away and if we've
already sent out a request to another TCU, we wait until that's
finished.

There is one thing to take care of: messages that have been sent out
cannot be aborted anymore, because they can only be sent exactly once
and they were successfully stored at the receiver side.

Note that we abort writes if they've been sent out, because the DATA
register might still hold something left to write. So, we could optimize
that more by letting it succeed if there is nothing more to write.

I've also decided to get rid of the dedicated ABORT register, but use a
privileged command for that purpose. In this way, we avoid having yet
another register with side effects and it fits better into the existing
concepts. PEMux will be the only component using it anyway.

I also revived the abort test to make sure that it works properly by
trying to abort SEND, REPLY, READ, and WRITE in every cycle during their
execution. The output of the abort test is constructed to be used within
our test system.