TCU: use BitUnion for EP registers.
This replaces bit fiddling at multiple places with a single definition
of the bit layout in the BitUnion. Additionally, we avoid the conversion
between the actual registers and our internal representation by simply
accessing the actual registers directly. And finally, we have a much
simpler and more reliable check for invalid EPs.