TCU: put registers into clearly defined groups.
We now have external registers, privileged registers, and unprivileged
registers. External registers are only writable externally, that is, via
the NoC. Privileged registers are intended for the privileged software
(PEMux). The TCU cannot enforce that these are only used by privileged
software though, but assumes that the privileged software maps this MMIO
region accordingly. Finally, unprivileged registers can be used by the
unprivileged software (again, not enforced by the TCU) and contain
primarily the registers to execute commands. Additionally, it contains
the CUR_TIME register.

As before, we have endpoint registers, which are only writable
externally and by privileged TCUs.