TCU: make remote memory accesses always refer to physical memory.
Otherwise, we need address translations on the remote PE, potentially
requiring a core request. The core or the software on the core could
decide to just not handle the core request and therefore keep the remote
memory access running forever.

Since we wanted a simple abort mechanism, abort is just waiting on the
remote PE until the memory access completed. And since an unabortable
command is currently running on our PE, we cannot even inform somebody
else to resolve this situation. Thus, without a more complicated abort
mechanism involving the remote PE (leading to new problems), remote
memory accesses to virtual memory require us to trust the software/core
in the remote PE.

For that reason, I decided to not allow that anymore, but allow remote
memory accesses only to physical memory. In this way, we can handle them
without core request, leading to the desired trust model.