TCU: make sure that the translation is gone before finishing the event.
Without this change it could happen that the a transfer was aborted and
therefore stopped after receiving all memory responses, but a
translation was still ongoing, because it was started after the abort.
Since we know that the translation can only still exists when the
transfer is stopped if it was aborted, we can safely abort the
translation and set it to NULL in this case.