TCU: don't undo credit reduction on failed sends and ACK immediately.
To simplify the commands as much as possible, I've now removed the
"cleanup actions" after a successful/failed SEND/REPLY. That is, we do
not increase the credits again if a SEND/REPLY has failed, which
originate from the message forwarding concept we had before. Also, we
immediately ACK the message on REPLY before sending the message out.