TCU: introduce EpFile as a preparation for in-memory EPs.
The EpFile allows the creation of EpCache's, which in turn can be used
to fetch and update EPs. Both actions are performed asynchronously and
call a registered callback on completion. With this, we are prepared for
in-memory EPs, which force us to send a memory request and thus, make EP
accesses asynchronous.

Since we now work on copies of EPs, we can no longer access the same EP
simultaneously. For example, the core could perform a REPLY, while the
TCU is receiving a message on the same receive EP. This worked fine
before, because we took care that we always left an EP in a consistent
state and, e.g., set the bit in the occupied mask at the beginning, etc.
However, with in-memory EPs we inherently work with a copy, because the
actual EP is stored in a cache (or DRAM).

I've decided to solve this problem in a simple way by disallowing
concurrent access to EPs. That is, we cannot have a command and a
receive operation taking place simultaneously. This is implemented by
letting each EpCache acquire a lock before fetching the EPs and
releasing it after being done. In theory, we could use more fine-grained
locking by doing that on a per-EP basis, but this makes it more
complicated and we almost never have simultaneous message receptions and
commands, so that we don't gain that much. And even if that happens, it
is always just a couple of cycles until the current one is done.