TCU: added type field to CORE_REQ to distinguish different usages.
Previously, we couldn't distinguish responses from requests and were
assuming that hardware could simply detect writes to CORE_REQ. This is
not the case, which is why we now use a type field to distinguish both
different requests form each other and also requests from responses.
Additionally, we represent "do nothing" as type=0.