TCU: don't change TCU state on SEND/REPLY errors.
Unfortunately, this causes some complications, because at first, we need
to access the EPs for that and cannot keep the EPs locked for the whole
time. Thus, we need to fetch the EPs again and deal with the
asynchronicity this adds to the command-finish-procedure.

Second, we cannot simply perform the credit reduction after receiving
the ACK from the receiver, because the receiver might otherwise send
the reply faster than the TCU can send the ACK. This currently only
happens with accelerators, but is a general problem, because we would
notify the TCU before sending the ACK. For that reason, we reduce the
credits right away and undo that in case of an error from the receiver
later.